chore(admin): reorganized some admin css/js views
