+++ views/default/admin/plugins/categories.css
@@ -1,0 +1,23 @@
+.elgg-admin-plugins-categories {
+	display: flex;
+	flex-wrap: wrap;
+	flex-direction: row;
+	margin-top: 1rem;
+	grid-gap: 0.5rem;
+	
+	> a {
+		padding: 0.25rem 0.5rem;
+		background: #e6e6ea;
+		border-radius: 3px;
+		font-size:0.85rem;
+		color: #2d3047;
+		text-decoration: none;
+		
+		&.elgg-state-selected,
+		&:hover {
+			color: #fff;
+			background: #2d3047;
+			text-decoration: none;
+		}
+	}
+}
