+++ views/default/resources/admin/plugin_text_file.php
@@ -1,46 +1,31 @@
 <?php
 
-use Michelf\MarkdownExtra;
+use Elgg\Exceptions\Http\EntityNotFoundException;
+use Elgg\Exceptions\Http\PageNotFoundException;
 
 elgg_set_context('admin');
+elgg_unregister_external_file('css', 'elgg');
 
 $plugin_id = (string) elgg_extract('plugin_id', $vars);
 $plugin = elgg_get_plugin_from_id($plugin_id);
-
-$filename = elgg_extract('filename', $vars);
-
-elgg_unregister_external_file('css', 'elgg');
-
 if (!$plugin instanceof \ElggPlugin) {
-	$error = elgg_echo('admin:plugins:markdown:unknown_plugin');
-	$body = elgg_view_layout('admin', [
-		'content' => $error,
-		'title' => $error,
-	]);
-	echo elgg_view_page($error, $body, 'admin');
-	return;
+	throw new EntityNotFoundException(elgg_echo('admin:plugins:markdown:unknown_plugin'));
 }
 
 $file_contents = false;
+$filename = elgg_extract('filename', $vars);
 if (in_array($filename, \ElggPlugin::ADDITIONAL_TEXT_FILES)) {
 	$file_contents = file_get_contents($plugin->getPath() . $filename);
 }
 
 if (!$file_contents) {
-	$error = elgg_echo('admin:plugins:markdown:unknown_file');
-	$body = elgg_view_layout('admin', ['content' => $error, 'title' => $error]);
-	echo elgg_view_page($error, $body, 'admin');
-	return;
+	throw new PageNotFoundException(elgg_echo('admin:plugins:markdown:unknown_file'));
 }
 
 $title = $plugin->getDisplayName() . ": $filename";
 
-$text = MarkdownExtra::defaultTransform($file_contents);
-
 $body = elgg_view_layout('admin', [
-	// setting classes here because there's no way to pass classes
-	// to the layout
-	'content' => elgg_format_element('div', ['class' => 'elgg-markdown'], $text),
+	'content' => elgg_view('admin/plugins/markdown', ['value' => $file_contents]),
 	'title' => $title,
 	'filter_id' => 'admin/plugin_text_file',
 	'filter_value' => "{$plugin_id}/{$filename}",
